Step 2

+++ src/main.rs
@@ -1,6 +1,7 @@
 use std::fs;
 use std::fs::File;
 use std::io::{BufWriter, Write};
+use std::path::Path;
 
 fn pack_color_rgba(r: u8, g: u8, b: u8, a: u8) -> u32 {
     let b1 = r as u32;
@@ -10,6 +11,7 @@ fn pack_color_rgba(r: u8, g: u8, b: u8, a: u8) -> u32 {
     (b4 << 24) + (b3 << 16) + (b2 << 8) + b1
 }
 
+//rust does not have operator overloading
 fn pack_color_rgb(r: u8, g: u8, b: u8) -> u32 {
     pack_color_rgba(r, g, b, 255)
 }
@@ -44,25 +46,101 @@ fn drop_ppm_image(filename: &str, image: &Vec<u32>, w: usize, h: usize) -> std::
 
         output.write(&[r, g, b])?;
     }
-
     //output closes at end of scope
     Ok(())
 }
 
+fn draw_rectangle(
+    img: &mut Vec<u32>,
+    img_w: usize,
+    img_h: usize,
+    x: usize,
+    y: usize,
+    w: usize,
+    h: usize,
+    color: u32,
+) {
+    // assert!(img.len() == img_w * img_h);
+    //move these asserts out of the loop for performance
+    // assert!(x + w < img_w && y + h < img_h);
+
+    for i in 0..w {
+        for j in 0..h {
+            let cx = x + i;
+            let cy = y + j;
+
+            img[cx + cy * img_w] = color;
+        }
+    }
+}
+
 fn main() -> std::io::Result<()> {
     let filename = "./out.ppm";
-    fs::remove_file(filename)?;
+    if Path::new(filename).exists() {
+        fs::remove_file(filename)?;
+    }
 
     let win_w: usize = 512;
     let win_h: usize = 512;
     let mut framebuffer: Vec<u32> = vec![255; win_w * win_h];
 
+    let map_w: usize = 16;
+    let map_h: usize = 16;
+    let mut map: Vec<&str> = "0000222222220000\
+                              1              0\
+                              1      11111   0\
+                              1     0        0\
+                              0     0  1110000\
+                              0     3        0\
+                              0   10000      0\
+                              0   0   11100  0\
+                              0   0   0      0\
+                              0   0   1  00000\
+                              0       1      0\
+                              2       1      0\
+                              0       0      0\
+                              0 0000000      0\
+                              0              0\
+                              0002222222200000"
+        .split("")
+        .collect(); // our game map
+                    //strip out null chars
+    map.remove(map_w * map_h - 1);
+    map.remove(0);
+
+    // println!("{:?}", map);
+    assert_eq!(map.len(), map_w * map_h);
+
     for j in 0..win_h {
+        //fill screen with a color gradient
         for i in 0..win_w {
             let r: u8 = (255.0 * (j as f32) / (win_h as f32)) as u8;
             let g: u8 = (255.0 * (i as f32) / (win_w as f32)) as u8;
             let b: u8 = 0;
             framebuffer[i + j * win_w] = pack_color_rgb(r, g, b);
+        }
+    }
+
+    let rect_w = win_w / map_w;
+    let rect_h = win_h / map_h;
+    for j in 0..map_h {
+        //draw the map
+        for i in 0..map_w {
+            if map[i + j * map_w] == " " {
+                continue;
+            } //skip empty spaces
+            let rect_x = i * rect_w;
+            let rect_y = j * rect_h;
+            draw_rectangle(
+                &mut framebuffer,
+                win_w,
+                win_h,
+                rect_x,
+                rect_y,
+                rect_w,
+                rect_h,
+                pack_color_rgb(0, 255, 255),
+            );
         }
     }
